requirements.txt
# ===============================
# Web Framework
# ===============================
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ===============================
# Core configuration
# ===============================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===============================
# Database
# ===============================
pymongo==4.6.1

# ===============================
# Supabase (stable set)
# ===============================
supabase==2.3.4
httpx==0.25.2
postgrest==0.13.2

# ===============================
# PDF Processing
# ===============================
pypdf==3.17.4

# ===============================
# HTTP / External calls
# ===============================
requests==2.31.0

# ===============================
# Health & monitoring
# ===============================
aiohttp==3.9.1
psutil==5.9.6
